# Inline Class#new.