--- a/debug_counter.h
+++ b/debug_counter.h
@@ -316,6 +316,9 @@ RB_DEBUG_COUNTER(obj_imemo_callinfo)
 RB_DEBUG_COUNTER(obj_imemo_callcache)
 RB_DEBUG_COUNTER(obj_imemo_constcache)
 
+RB_DEBUG_COUNTER(opt_new_hit)
+RB_DEBUG_COUNTER(opt_new_miss)
+
 /* ar_table */
 RB_DEBUG_COUNTER(artable_hint_hit)
 RB_DEBUG_COUNTER(artable_hint_miss)