@k0kubun 2025-04-09
The YJIT implementation looks good.

As to the interpreter, I tried to come up with a simpler instruction sequence, but the current one seemed reasonable as is.