@tenderlove 2025-04-09
> As to the interpreter, I tried to come up with a simpler instruction sequence, but the current one seemed reasonable as is.

As [I mentioned on the redmine ticket](https://bugs.ruby-lang.org/issues/21254#note-2), I think we can remove one of the swap instructions, but it will take some work. I was able to remove it, but Coverage tests fail and I'm not sure how to fix them yet.