@tekknolagi 2025-04-25
You need to run zjit-bindgen because opt_new shifted some op numbers around.